+++ SharePay/Sourses/Purchase/Model/Entity.swift
@@ -13,12 +13,12 @@ struct Purchase{
     var name: String
     var emoji: String
     var draft: Bool
-    var id: Int64
-    var amount: Int64
+    var id: Int
+    var amount: Int
     var created_at: Date
     
-    init(id: Int64, name: String = "", description: String = "", emoji: String = "",
-         draft: Bool = true, amount: Int64 = 0, created_at: Date = Date.now){
+    init(id: Int, name: String = "", description: String = "", emoji: String = "",
+         draft: Bool = true, amount: Int = 0, created_at: Date = Date.now){
         self.id = id
         self.name =  name
         self.description = description
@@ -32,8 +32,8 @@ struct Purchase{
 struct PurchaseParticipant {
     var phoneNumber: String
     var name: String
-    var amount: Int64
-    init(phoneNumber: String, name: String, amount: Int64){
+    var amount: Int
+    init(phoneNumber: String, name: String, amount: Int){
         self.amount = amount
         self.name = name
         self.phoneNumber = phoneNumber
